requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-taggit==0.22.2
flake8
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.7
psycopg2
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
